Register client render hooks explicitly instead of auto-subscriber scanning
