+++ src/main/java/com/thunder/wildernessodysseyapi/client/ChipSetRendererRegistration.java
@@ -1,19 +1,13 @@
-package com.thunder.wildernessodysseyapi.Client;
+package com.thunder.wildernessodysseyapi.client;
 
-import com.thunder.wildernessodysseyapi.Core.ModConstants;
 import com.thunder.wildernessodysseyapi.item.ModItems;
-import net.neoforged.api.distmarker.Dist;
-import net.neoforged.bus.api.SubscribeEvent;
-import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
 import top.theillusivec4.curios.api.client.CuriosRendererRegistry;
 
-@EventBusSubscriber(modid = ModConstants.MOD_ID, value = Dist.CLIENT)
 public final class ChipSetRendererRegistration {
     private ChipSetRendererRegistration() {
     }
 
-    @SubscribeEvent
     public static void onClientSetup(FMLClientSetupEvent event) {
         event.enqueueWork(() -> CuriosRendererRegistry.register(ModItems.CLOAK_CHIP.get(), ChipSetCurioRenderer::new));
     }
