+++ src/main/java/com/thunder/wildernessodysseyapi/client/NeuralFrameLayerRegistration.java
@@ -1,19 +1,17 @@
-package com.thunder.wildernessodysseyapi.Client;
+package com.thunder.wildernessodysseyapi.client;
 
-import com.thunder.wildernessodysseyapi.Core.ModConstants;
 import net.minecraft.client.renderer.entity.player.PlayerRenderer;
 import net.minecraft.client.resources.PlayerSkin;
-import net.neoforged.api.distmarker.Dist;
-import net.neoforged.bus.api.SubscribeEvent;
-import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.client.event.EntityRenderersEvent;
 
-@EventBusSubscriber(modid = ModConstants.MOD_ID, value = Dist.CLIENT)
-public final class CloakLayerRegistration {
-    private CloakLayerRegistration() {
+public final class NeuralFrameLayerRegistration {
+    private NeuralFrameLayerRegistration() {
     }
 
-    @SubscribeEvent
+    public static void onRegisterLayerDefinitions(EntityRenderersEvent.RegisterLayerDefinitions event) {
+        event.registerLayerDefinition(NeuralFrameModel.LAYER_LOCATION, NeuralFrameModel::createLayer);
+    }
+
     public static void onAddLayers(EntityRenderersEvent.AddLayers event) {
         for (PlayerSkin.Model skin : event.getSkins()) {
             addLayer(event, skin);
@@ -23,7 +21,7 @@ public final class CloakLayerRegistration {
     private static void addLayer(EntityRenderersEvent.AddLayers event, PlayerSkin.Model skin) {
         PlayerRenderer renderer = event.getSkin(skin);
         if (renderer != null) {
-            renderer.addLayer(new CloakRenderLayer(renderer));
+            renderer.addLayer(new NeuralFrameRenderLayer(renderer));
         }
     }
 }
