+++ src/main/java/com/thunder/wildernessodysseyapi/client/weather/PurpleStormClientEffects.java
@@ -1,4 +1,4 @@
-package com.thunder.wildernessodysseyapi.Client.weather;
+package com.thunder.wildernessodysseyapi.client.weather;
 
 import com.thunder.wildernessodysseyapi.Core.ModConstants;
 import net.minecraft.client.Minecraft;
